+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fixed `array[idx].as<JsonVariant>()` and `object[key].as<JsonVariant>()`
 * Fixed return value of `JsonObject::set()` (issue #350)
+* Fixed undefined behavior in `Prettyfier` and `Print` (issue #354)
 
 v5.6.6
 ------